Solves a bug in tailRec that, by using Sets, gets confused by nodes that are equal in the tree.
The idea is to use another Stack for the expanded nodes and, by construction, when we arrive for the second time to the same node, if this has already been expanded, it will be at the top of the expanded Stack. That is what allows us to evaluate the visited condition without any traversal. And, obviously, when we have detected that the node has been visited, not only we push the tree formed by the two top elements of the done stack, but we pop the visited node from the visited stack.